moved browser_setup to scraping
