+++ src/energymonitor/__init__.py
@@ -1,0 +1,1 @@
+from energymonitor.scraping import browser_setup
